@@ -35,56 +35,31 @@
     "prepush": "npm run -s check",
     "preparecommitmsg": "prepare-commit-msg-angular $2 $3"
   },
-  "nyc": {
-    "reporter": [
-      "lcov",
-      "text"
-    ],
-    "all": true,
-    "require": [
-      "babel-register"
-    ],
-    "include": [
-      "src/"
-    ],
-    "exclude": [
-      "dist/**",
-      "node_modules/**",
-      "docs/**",
-      "test/**"
-    ],
-    "sourceMap": false,
-    "instrument": false,
-    "check-coverage": true,
-    "lines": 100,
-    "functions": 100,
-    "branches": 100,
-    "statements": 100
-  },
   "dependencies": {
-    "react": "^15.0.0",
-    "react-dom": "^15.0.0"
+    "react": "^15.4.2",
+    "react-dom": "^15.4.2"
   },
   "devDependencies": {
-    "babel-cli": "^6.18.0",
-    "babel-core": "^6.18.2",
-    "babel-plugin-istanbul": "^3.0.0",
-    "babel-preset-latest": "^6.16.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-2": "^6.18.0",
-    "babel-register": "^6.18.0",
+    "babel-cli": "^6.24.0",
+    "babel-core": "^6.24.0",
+    "babel-plugin-istanbul": "^4.1.1",
+    "babel-preset-env": "^1.3.2",
+    "babel-preset-react": "^6.23.0",
+    "babel-preset-stage-2": "^6.22.0",
+    "babel-register": "^6.24.0",
     "chai": "^3.5.0",
     "chai-enzyme": "^0.6.1",
-    "coveralls": "^2.11.15",
+    "coveralls": "^2.13.0",
     "dirty-chai": "^1.2.2",
-    "enzyme": "^2.6.0",
+    "enzyme": "^2.8.0",
+    "eslint": "^3.19.0",
     "eslint-config-fernandopasik": "^0.5.1",
-    "husky": "^0.11.9",
+    "husky": "^0.13.3",
     "mocha": "^3.2.0",
-    "nyc": "^10.0.0",
+    "nyc": "^10.2.0",
     "prepare-commit-msg-angular": "^0.2.0",
-    "react-addons-test-utils": "^15.0.0",
-    "validate-commit-msg": "^2.8.2"
+    "react-addons-test-utils": "^15.4.2",
+    "validate-commit-msg": "^2.12.0"
   },
   "engines": {
     "node": ">=6.0.0",
