@@ -1,9 +1,6 @@
-require('babel-register')({ plugins: [ 'istanbul' ] });
-
-const
-  chai = require('chai'),
-  chaiEnzyme = require('chai-enzyme'),
-  dirtyChai = require('dirty-chai');
+import chai from 'chai';
+import chaiEnzyme from 'chai-enzyme';
+import dirtyChai from 'dirty-chai';
 
 global.expect = chai.expect;
 chai.should();
